Bug 902070 - [email/activesync] Change jswbxml to run its unit tests under mocha. r=squib
Notable things:

- Our node.js version has some DOM deps we need to load in.  It's conceivable I
actually broke something with running this in the web browser.  Those should be
minor and I'll make any minor changes and rebase them when running the GELAM
tests which would notice this breakage.

- TextEncoder can no longer produce us-ascii output, so I made all the tests
use utf-8.  It's possible we should actually keep using the polyfill that
is fine with creating us-ascii output, although that would be somewhat
unrealistic.  It seems like if we really want us-ascii (even though the tests
didn't break?) that we should probably do something special for that.